Alteração na função sum

+++ GitHub1/Source.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-double sum(double a, double b)
+int sum(int a, int b)
 {
 	return a + b;
 }
